Employ Test Run's wiki implement SimpleTestRun
Each Polarion's Test Run has a wiki page which is a typed
text so we can use it to implement SimpleTestRun.
(Originally I proposed Test Run's description for that
- that would be even worse...) Hence SimpleTestRun and
SimpleTestPlan are mapped nearly the same way to their
Polarion counterparts.

Anyway, we are talking about a temporary implementation.
Both SimpleTestRun and SimpleTestPlan should be mapped
more naturally and safely, probably using new custom
fields for Test Runs and Documents.