work_item import moved to global namespace
Because of circular references, work_item had been imported in a class function
instead of in the global namespace. This caused it to be hidden from the rest
of the module. Becuase plan is only referenced in a single place in work_item,
I was able to change the references so that work_item would be imported into
the global namespace

Change-Id: Icecb0a143a1f2337c8eecac6185f18b1957fd362