obj_setter id_field fix
The obj_setter had assumed that the first parameter was the id
In _WorkItems the first field is the project_id so the class
was failing on instantiation. This fix puts the id_field into
the parameter dict and removes the assumption.

Removed required fields from TestRecord, when I saw a use case
of adding TestRecords before they are executed.

Removed use of the required fields in the add test record
method of the TestRun

Change-Id: I5e9555459f5427ac50642ac895012c8b5ffd16c4