enums and  id attribute name
All the Enum classes were erased as they are not used for anything.

The EnumOptionId and CustomFieldType objects both used id as their
key field, which is a reserved keyword. This causes a the namespace
to be overwritten when using the id field as an argument in the object
constructor. EnumCustomFieldType has the same id as CustomFieldType

Change-Id: I52aa1e59a7357e6cb2b1fbbf2a2bd2f223fd4c60