URI convert
Get the ID field straight from the URI instead of instantiating an
object. This saves a lot of time. For setting, it still instatiates
the object, so it can verify that the id is correct.

The Project and Document objects have slightly different URI structs
so their struct is defined in their class.

Change-Id: Ie6dec4e919af6f81213368304542333d1cf933f0