modify workitem query function.
Until now the query function worked across projects and work item types.
This fix changes the functionality so that for each workitem type, it
only queries either the passed in project or the default project and
only the specified work item type.
To use the original functionality, the parent _WorkItem class function
can still be used.

Change-Id: Ib7cc60a19d94f72fa4588ede4eea2fc011af85e7