TestRun add record performance patch
This patch increases the performance of adding a new test record to
a test run. The client has to enforce the business logic that does
not allow the same test case to be entered into a run multiple times.
The original code dowloaded the TestRun object and checked if the test
record was contained in it. This took a lot of time when there were a
lot of test records. This patch changes the method and instead calls the
search function to check if the test record exists in the TestRun.
This takes less time and should take approximately the same amount of
time irrespective of the number of test records in the run.

Change-Id: I2885a72d3e610ea13474960c80944b22c243c2d2