Fixed: '#392303: pylarion bug - test run planned in field'
We use '__metaclass__' to wrap pylarion functions, as a result
the metadata of a wrapped function is the wrapper metadata.
So when inspecting 'Plan', which is 'TestRun' 'plannedin' field,
the inspected arguments are belong to the wrapper.
I used 'decorator' package to fixed this problem.
Also added new Tests and 'decorator' to 'requirements.txt'

fixed minor bug in base_polarion._obj_setter whereby a None value
raised an error.

Change-Id: I6985df474583d80e9deb69929a87680cdf2ec3be