cachingpolicy and timeout options
Added cachingpolicy and timeout options to the configuration file.

cachingpolicy is a configuration parameter that specifies either 0 or 1.
When it is set correctly, the client always goes to the same host when using
a load balancer. Unfortunately, some workstations require 1 and others 0 and
I have not understood what the qualification is. (1 or 0, default 0)

timeout is the HTTP tiomeout. (120 secs default)

Change-Id: I40822702421bbf004c69b5e386253c53648203eb