Bug-fix(POLARION-974): Using required fields from TestRun template
Bug descreption: When the user uses 'TestRun.create' he needs to
add the required fields to 'kwargs', this is a duplicate effort
in the case where these fields are already set in the template.

I added a 'tx_wrapper' to 'TestRun.create' so we first create the
TestRun with the template, and then if the required fields are
missing we preform a rollback and throw exception.

Change-Id: I6eb2c28f9be10c2c46a20942490bb34beb9c32ea