Added stickiness to node when using load balancer
There is a bug that in certain undefined circumstances the WSDL will
switch nodes when using a load balancer. This causes the error "Not Authorized"
because the login in Polarion is per node. This fix takes the specific node
logged into and makes that the url that all services should log into. This
removes the problem. Other options tried were to pass in the load balancer's
sticky bit cookie. Though I tried numerous ways, this did not solve the problem

Change-Id: I0a189500bac520d3614d09288404e42395511974