After much research discovered that 'httplib' doesn't verify certificate by default and when being told to do so it requires the location of the CA-bundle. Also discovered that 'requests' lib also requires the location of the CA-bundle.
The way to solve it is to define ENV-Variabels that these libaries use
to get the path for the system CA-bundle.

Because of the fact that the user will have to define, in one way
or another, the path to the CA-bundle, I decided to add additonal
line to pylarion.cfg which specify the path to the CA-bundle.

By default it will be commented-out and if the user will want
certificate verification he will have to uncomment this line
and verify the path.

Also made all necessery changes to support this.

Change-Id: I286a45bdceba7342242069191c0551c145d6649f