Auto generate work items
Work items are configured per instance, so the workitems module was
modified to generate the workitems on the fly. To do this the following
changes were made:
pylarion.cfg - added workitems option. This is a JSON detailing the
    workitems you want to create. If this is blank, it will connect
    to the server to get the list.
base_polarion.py - Moved the Configuration to its own class. This allows
    it to be accessed from other modules (such as workitems).
    Changed the control element of the csm to use the _wi_type.
work_item.py - Removed all the individual workitem classes.
    Removed the control element of the csm for custom enums.
    Added code to create the workitems.
README.txt and index.rst - included these elements in the documentation.

Change-Id: Ifd26d8a8a863c29fe43cd873c382427fc2552620