removed logstash functionality
The logstash functionality gave us usage logs on the entire library.
Unfortunately, our server stopped working and users are complaining
that it harms performance. So I removed it from the code. in the future
I may add parallel processing so this isn't a problem..

tests were changed for new ci env

Change-Id: Iffd4f6b44c7c68acfae4f74f9e9ed1024066e8ea