improve pylarion-cmd
Add the feature of creating document in polarion from external file or
text. It's for solving the problem of importing document to polarion.
The document with HTML markup can be all supported.

One scenario:
Importing a google (or LibreOffice) document to polarion via command line.
Here are two steps:
1) download Gdoc as .html file
2) use pylaron-cmd to import it polarion

For example:
$ pylarion-cmd update --is_document --doc_with_space='KernelStorageQE/DOC-17'
  --doc_content='/tmp/gdoc.html'
  Created document 'KernelStorageQE/DOC-17'
   - document author      : xhe
   - document type        : generic
   - allowed workitem type: testcase
   - created date         : 2018-04-24 11:25:10.000761
   - document content is from: /root/my_gdoc1.html

Change-Id: I5b7f9ff684dcf19c4be75092c7ed5c0a6ebab2f9
Signed-off-by: Xiaonan He <xhe@redhat.com>