fix: restore dynamic imports removed by ruff modernization
The ArrayOfEnumOptionId import was accidentally removed in commit
992340c when ruff consolidated imports. The original # NOQA comment
was not recognized by ruff, causing the "unused" import to be removed.

This import is required for dynamic custom field type resolution in
get_custom_fields() which uses globals()[type_name] to look up classes.

Also adds Properties import which may be needed for other custom field
types.

Signed-off-by: Wayne Sun <gsun@redhat.com>